🏗️ Random dots sort of works. Needs tweaking
Need to tweak it so that the dots are random, and not in a line (left/right).

Ideally, and this is for the future - you should be able to spread these dots up/down/left/right, not just in a line left to right until you hit the end of the row.